WorkingOnLighttest
We need to modify Dataloader to both load train and test data and give labels to test data. Then we need to make sure testfunction works. Then we should try to run light test, to train networks with x lightsources and test them and then plot a linear graph of the accuracy and other metrics returned as result of the test function.